US122816 - Refactor Activities attachment file picker to support entity and link
